+++ src/http/register-websocket/connection.js
@@ -3,7 +3,9 @@ let pool = require('./pool')
 let noop = err => err ? console.log(err) : ''
 let { updater } = require('@architect/utils')
 
-module.exports = function connection ({ get }, connectionId, ws) {
+module.exports = function connection (inventory, connectionId, ws) {
+  let { get } = inventory
+
   // Save this for send to use
   pool.register(connectionId, ws)
   let update = updater('Sandbox')
@@ -16,7 +18,8 @@ module.exports = function connection ({ get }, connectionId, ws) {
       invoke({
         lambda,
         body: msg,
-        connectionId
+        connectionId,
+        inventory,
       }, noop)
     }
     else {
@@ -25,7 +28,8 @@ module.exports = function connection ({ get }, connectionId, ws) {
       invoke({
         lambda,
         body: msg,
-        connectionId
+        connectionId,
+        inventory,
       }, noop)
     }
   })
@@ -36,7 +40,8 @@ module.exports = function connection ({ get }, connectionId, ws) {
     invoke({
       lambda,
       connectionId,
-      req: { headers: { host: `localhost:${process.env.PORT}` } }
+      req: { headers: { host: `localhost:${process.env.PORT}` } },
+      inventory,
     }, noop)
   })
 }