+++ src/http/register-websocket/upgrade.js
@@ -8,7 +8,9 @@ let { updater } = require('@architect/utils')
  * - In APIGWv2, !2xx responses hang up and return the status code
  * - However, 2xx responses initiate a socket connection (automatically responding with 101)
  */
-module.exports = function upgrade (wss, { get }) {
+module.exports = function upgrade (wss, inventory) {
+  let { get } = inventory
+
   return function upgrade (req, socket, head) {
 
     // Get the $connect Lambda
@@ -23,7 +25,8 @@ module.exports = function upgrade (wss, { get }) {
     invoke({
       lambda,
       connectionId,
-      req
+      req,
+      inventory,
     },
     function connect (err, res) {
       let statusCode = res && res.statusCode