+++ test/mock/dep-warn/src/http/get-deps_in_shared/index.js
@@ -1,0 +1,9 @@
+let lambda = require('lambda-dep')
+let shared = require('@architect/shared')
+
+exports.handler = async function http (req) {
+  return {
+    statusCode: 200,
+    body: 'Henlo'
+  }
+}