+++ test/mock/dep-warn/src/shared/index.js
@@ -1,0 +1,3 @@
+let lambda = require('lambda-dep')
+
+exports.handler = lambda